pyproject.toml
[project]
name = "pywebview5-4-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pywebview==5.4",
]
